♻️ refactor: token 검증 메서드 생성
